Improved the logic used to determine when to create _cor files in the drizCR step.
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/betadrizzle@13419 fe389314-cf27-0410-b35b-8c050e845b92

+++ lib/betadrizzle/drizCR.py
@@ -282,7 +282,7 @@ def _drizCr(sciImage,paramDict,saveFile=True):
             __corrFile = np.zeros(__inputImage.shape,dtype=__inputImage.dtype)
             __corrFile = np.where(np.equal(__dqMask,0),__blotData,__inputImage)
 
-            if(saveFile):
+            if(saveFile and paramDict['driz_cr_corr']):
                 # Remove the existing cor file if it exists
                 if(os.access(crCorImage, os.F_OK)):
                     os.remove(crCorImage)
